requirements/base.txt
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-environ==0.8.1
drf-yasg==1.20.0
git+https://github.com/schecterdamien/django-request-context.git
django-model-utils==4.2.0
python-dateutil==2.8.2
django-cors-headers==3.12.0
psycopg2-binary==2.9.3
#psycopg2==2.9.3
gunicorn==20.1.0
gevent==21.12.0
django-error-report==0.2.0
django-q==1.3.9
redis==3.5.3
Fabric==2.7.0
playwright==1.22.0
lxml==4.9.0
dateparser==1.1.1
price-parser==0.3.4
